Adds copy of pantry switch to recipe generator
Because visibility of program state is important! Also, I ran Prettier, so the code looks way different, but I essentially just copied and adjusted some of the code from VirtualPantry.jsx to fit this. The functions of note are the extra useStates (pantries and pantryName) initial useEffect, the function fetchPantries, and the extra form field.