Adds env file reading to backend
Add the file keys.env in the same directory as the backend index.js, then add your OpenAI key and MongoDB URL as environment variables named as used in index.js. I was simply getting annoyed at having to replace the urls every time. There are technically benefits in terms of being able to use different keys for deployment, but for now, it's just for convenience.